requirements.txt
aiohttp==3.8.5
allure-subtests==0.1.9
pytest-check==2.4.1
allure_pytest==2.13.5
allure_python_commons==2.13.5
beautifulsoup4==4.12.3
lxml==4.9.3
pytest==8.3.2
pytest-aiohttp==1.0.5